chore(spanner): Fix rebase issues
